+++ airline-core/src/main/moditect/module-info.java
@@ -18,7 +18,7 @@ module com.github.rvesse.airline
   requires com.github.rvesse.airline.io;
   requires org.apache.commons.lang3;
   requires org.apache.commons.collections4;
-  requires javax.inject;
+  requires java.inject;
 
   exports com.github.rvesse.airline;
   exports com.github.rvesse.airline.annotations;