+++ airline-core/src/test/java/com/github/rvesse/airline/parser/TestOptionParsing.java
@@ -22,7 +22,9 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import com.github.rvesse.airline.CommandLineInterface;
+import com.github.rvesse.airline.annotations.Arguments;
 import com.github.rvesse.airline.annotations.Command;
+import com.github.rvesse.airline.annotations.DefaultOption;
 import com.github.rvesse.airline.annotations.Option;
 import com.github.rvesse.airline.builder.CliBuilder;
 import com.github.rvesse.airline.parser.errors.ParseArgumentsUnexpectedException;
@@ -32,6 +34,7 @@ import com.github.rvesse.airline.parser.options.AbstractNameValueOptionParser;
 import com.github.rvesse.airline.parser.options.ClassicGetOptParser;
 import com.github.rvesse.airline.parser.options.ListValueOptionParser;
 import com.github.rvesse.airline.parser.options.LongGetOptParser;
+import com.github.rvesse.airline.parser.options.MaybeListValueOptionParser;
 import com.github.rvesse.airline.parser.options.MaybePairValueOptionParser;
 import com.github.rvesse.airline.parser.options.StandardOptionParser;
 
@@ -41,13 +44,28 @@ public class TestOptionParsing {
     public static class OptionParsing {
 
         @Option(name = { "-a", "--alpha" })
-        private boolean alpha;
+        boolean alpha;
 
         @Option(name = { "-b", "--beta" }, arity = 1)
-        private String beta;
+        String beta;
 
-        @Option(name = { "-c", "--charlie" }, arity = 2)
-        private List<String> charlie = new ArrayList<String>();
+        @Option(name = { "-c", "--charlie" }, arity = 2) 
+        List<String> charlie = new ArrayList<String>();
+    }
+    
+    @Command(name = "OptionParsing1")
+    public static class OptionAndArgumentParsing extends OptionParsing {
+        
+        @Arguments
+        List<String> args;
+    }
+    
+    @Command(name = "OptionParsing1")
+    public static class OptionAndDefaultParsing extends OptionParsing {
+        
+        @Option(name = { "-d", "--delta" }, arity = 1)
+        @DefaultOption
+        String delta;
     }
 
     private <T> T testParsing(CommandLineInterface<T> parser, String... args) {
@@ -353,6 +371,236 @@ public class TestOptionParsing {
                .withOptionParser(new ListValueOptionParser<T>(listSeparator));
         //@formatter:on
         return builder.build();
+    }
+    
+    private <T> CommandLineInterface<T> createMaybeListValueParser(Class<? extends T> cls, char listSeparator) {
+        //@formatter:off
+        CliBuilder<T> builder = CommandLineInterface.<T>builder("test")
+                                   .withCommand(cls);
+        builder.withParser()
+               .withOptionParser(new MaybeListValueOptionParser<T>(listSeparator));
+        //@formatter:on
+        return builder.build();
+    }
+    
+    @Test(expectedExceptions = ParseOptionMissingValueException.class, expectedExceptionsMessageRegExp = "Required values.*")
+    public void option_parsing_maybe_list_value_bad_01() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-c");
+    }
+
+    @Test(expectedExceptions = ParseOptionMissingValueException.class, expectedExceptionsMessageRegExp = "Too few.*")
+    public void option_parsing_maybe_list_value_bad_02() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-c", "one");
+    }
+    
+    @Test(expectedExceptions = ParseOptionUnexpectedException.class, expectedExceptionsMessageRegExp = "Too many.*")
+    public void option_parsing_maybe_list_value_bad_03() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-c", "one,two,three");
+    }
+    
+    @Test(expectedExceptions = ParseOptionMissingValueException.class, expectedExceptionsMessageRegExp = "Too few.*")
+    public void option_parsing_maybe_list_value_bad_04() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-cone");
+    }
+    
+    @Test(expectedExceptions = ParseOptionUnexpectedException.class, expectedExceptionsMessageRegExp = "Too many.*")
+    public void option_parsing_maybe_list_value_bad_05() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-cone,two,three");
+    }
+    
+    @Test(expectedExceptions = ParseOptionUnexpectedException.class, expectedExceptionsMessageRegExp = "Too many.*")
+    public void option_parsing_maybe_list_value_bad_06() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-c", "one", "two", "three");
+    }
+    
+    @Test(expectedExceptions = ParseOptionUnexpectedException.class, expectedExceptionsMessageRegExp = "Too many.*")
+    public void option_parsing_maybe_list_value_bad_07() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-c", "one", "two", "three", "-a");
+    }
+    
+    @Test(expectedExceptions = ParseOptionUnexpectedException.class, expectedExceptionsMessageRegExp = "Too many.*")
+    public void option_parsing_maybe_list_value_bad_08() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-c", "one", "two", "three", "-afoo");
+    }
+    
+    @Test(expectedExceptions = ParseOptionUnexpectedException.class, expectedExceptionsMessageRegExp = "Too many.*")
+    public void option_parsing_maybe_list_value_bad_09() {
+        CommandLineInterface<OptionAndArgumentParsing> parser = createMaybeListValueParser(OptionAndArgumentParsing.class, ',');
+        testParsing(parser, "OptionParsing1", "-c", "one", "two,three", "--", "foo");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_01() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-b", "foo");
+
+        Assert.assertEquals(cmd.beta, "foo");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_02() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-bfoo");
+
+        Assert.assertEquals(cmd.beta, "foo");
+    }
+
+    @Test
+    public void option_parsing_maybe_list_value_03() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one,two");
+
+        Assert.assertEquals(cmd.charlie.size(), 2);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_04() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-cone,two");
+
+        Assert.assertEquals(cmd.charlie.size(), 2);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_05() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-b", "foo,bar");
+
+        Assert.assertEquals(cmd.beta, "bar");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_06() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-cone,two,three,four");
+
+        Assert.assertEquals(cmd.charlie.size(), 4);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.charlie.get(2), "three");
+        Assert.assertEquals(cmd.charlie.get(3), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_07() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-cone", "two", "three,four");
+
+        Assert.assertEquals(cmd.charlie.size(), 4);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.charlie.get(2), "three");
+        Assert.assertEquals(cmd.charlie.get(3), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_08() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one,two", "three,four");
+
+        Assert.assertEquals(cmd.charlie.size(), 4);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.charlie.get(2), "three");
+        Assert.assertEquals(cmd.charlie.get(3), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_09() {
+        CommandLineInterface<OptionParsing> parser = createMaybeListValueParser(OptionParsing.class, ',');
+        OptionParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one,two", "three", "four");
+
+        Assert.assertEquals(cmd.charlie.size(), 4);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.charlie.get(2), "three");
+        Assert.assertEquals(cmd.charlie.get(3), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_10() {
+        CommandLineInterface<OptionAndArgumentParsing> parser = createMaybeListValueParser(OptionAndArgumentParsing.class, ',');
+        OptionAndArgumentParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one,two", "three", "four");
+
+        Assert.assertEquals(cmd.charlie.size(), 2);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.args.size(), 2);
+        Assert.assertEquals(cmd.args.get(0), "three");
+        Assert.assertEquals(cmd.args.get(1), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_11() {
+        CommandLineInterface<OptionAndArgumentParsing> parser = createMaybeListValueParser(OptionAndArgumentParsing.class, ',');
+        OptionAndArgumentParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one", "two,three", "four");
+
+        Assert.assertEquals(cmd.charlie.size(), 4);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.charlie.get(2), "three");
+        Assert.assertEquals(cmd.charlie.get(3), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_12() {
+        CommandLineInterface<OptionAndArgumentParsing> parser = createMaybeListValueParser(OptionAndArgumentParsing.class, ',');
+        OptionAndArgumentParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one", "two", "three", "four");
+
+        Assert.assertEquals(cmd.charlie.size(), 2);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.args.size(), 2);
+        Assert.assertEquals(cmd.args.get(0), "three");
+        Assert.assertEquals(cmd.args.get(1), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_13() {
+        CommandLineInterface<OptionAndArgumentParsing> parser = createMaybeListValueParser(OptionAndArgumentParsing.class, ',');
+        OptionAndArgumentParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one", "two", "--", "three", "four");
+
+        Assert.assertEquals(cmd.charlie.size(), 2);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.args.size(), 2);
+        Assert.assertEquals(cmd.args.get(0), "three");
+        Assert.assertEquals(cmd.args.get(1), "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_14() {
+        CommandLineInterface<OptionAndDefaultParsing> parser = createMaybeListValueParser(OptionAndDefaultParsing.class, ',');
+        OptionAndDefaultParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one", "two", "three", "four");
+
+        Assert.assertEquals(cmd.charlie.size(), 2);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.delta, "four");
+    }
+    
+    @Test
+    public void option_parsing_maybe_list_value_15() {
+        CommandLineInterface<OptionAndDefaultParsing> parser = createMaybeListValueParser(OptionAndDefaultParsing.class, ',');
+        OptionAndDefaultParsing cmd = testParsing(parser, "OptionParsing1", "-c", "one", "two", "-b", "three", "four");
+
+        Assert.assertEquals(cmd.charlie.size(), 2);
+        Assert.assertEquals(cmd.charlie.get(0), "one");
+        Assert.assertEquals(cmd.charlie.get(1), "two");
+        Assert.assertEquals(cmd.beta, "three");
+        Assert.assertEquals(cmd.delta, "four");
     }
     
     @Test