+++ CHANGELOG.md
@@ -12,13 +12,38 @@ API refactoring based on several years of deployed production usage and communit
     - New `CommandContext` abstraction in `CommandFactory` API
     - New `airline-guice` module for Google Guice integration
 
----
+
 
 # 2.x Releases
+=======
 
-Major refactoring and functionality enhancements
+## 2.8.2
+
+
+## 2.8.1
+
+- Bug Fixes
+    - Fix for using Airline in JPMS environments (#102) (thanks to @jfallows)
 
 ## 2.8.0
+
+- Core Improvements
+    - `@Option` annotation now allows an array of values for the `title` field allowing options with `arity` > 1 to have distinct titles shown for each value the option accepts
+
+
+## 2.7.2
+
+- Bug Fixes
+    - `@AllowedEnumValues` incorrectly provides help hint that values are case insensitive when they are actually case sensitive (#100)
+    - `@AllowedRawValues` does not handle case insensitive mode correctly (#100)
+    - `@Option` with `type = OptionType.GLOBAL` and `type = OptionType.GROUP` are not correctly parsed at the command level (#101)
+
+## 2.7.1
+
+- Bug Fixes
+    - `CollectAll` handler collected multiple duplicates of the same error when certain restrictions were used.  Now instead duplicates are combined into a single exception (#94)
+    - CLIs could allow command classes with the same declared command name in their `@Command` annotation to silently override each other.  Now instead an error will be thrown reporting the problem (#95)
+    - When using group default commands and command abbreviation we can incorrectly ignore a value (#99)
 
 ## 2.7.0